Update website to use Altserver and add note
SideServer has been causing issues with SideStore installs and adding these instructions in place will help with support on the discord and will not have to redirect instructions to the wiki. unfortunately I couldn't figure how to just add them as a default instructions and have SideServer instructions as a alternative method.

Signed-off-by: Spidy123222 <64176728+Spidy123222@users.noreply.github.com>